change the pointee of `sscanf` `%n` to `int` (not `size_t`)

+++ lib/orion3-config-parser/parser.c
@@ -28,7 +28,7 @@ orion3_param* orion3_param_read_input_file(char* filename) {
 		/* read in each float out of the buffer */
 		char* sscanf_buffer = buffer;
 		int sscanf_count = -1;
-		size_t buffer_offset = 0;
+		int buffer_offset = 0; /* not size_t, used in %n */
 		float single_scale = 0.0;
 		while( sscanf_count = sscanf(sscanf_buffer, "%f%n", &single_scale, &buffer_offset) ) {
 			if( sscanf_count < 1 )
